\name{EAtestslite-package}
\alias{EAtestslite-package}
\alias{EAtestslite}
\docType{package}
\title{
\packageTitle{EAtestslite}
}
\description{
\packageDescription{EAtestslite}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EAtestslite}
\packageIndices{EAtestslite}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{EAtestslite}

Maintainer: \packageMaintainer{EAtestslite}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
